+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `DAVESession.epoch`
+- `DAVESession.ownLeafIndex`
+
+### Changed
+
+- Bumped Node-API version to 6
+
 ## [0.1.2] - 2025-04-04
 
 ### Added